requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python-headless
streamlit-drawable-canvas
